requirements.txt
django==2.2.3
supervisor==4.1.0
# simpleui
paramiko==2.6.0
channels==2.2.0
channels-redis==2.4.0
celery==4.3.0
redis==3.3.4
eventlet==0.23.0
selectors2==2.0.1
django-redis==4.10.0
pyguacamole==0.8
ansible==2.8.5
daphne==2.3.0
gunicorn==19.9.0
gevent==1.4.0
async==0.6.2
pymysql==0.9.3
django-ratelimit==2.0.0
cryptography==2.7
# mysqlclient 比 pymysql 性能高，安装前需要：
# apt-get install python3-dev default-libmysqlclient-dev        # debian / Ubuntu
# yum install python3-devel mysql-devel     # Red Hat / CentOS
# mysqlclient==1.4.4
